refactor(admin): extract site replication identity helpers
